requirements.txt
matplotlib==3.1.1
numpy==1.17.2
opencv-python==4.1.0.25
pandas==0.24.2
pyssim==0.4
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.2.2